+++ hs-bindgen-runtime/test/test-runtime.hs
@@ -6,7 +6,8 @@ import Test.HsBindgen.Runtime.Bitfield qualified
 import Test.HsBindgen.Runtime.CBool qualified
 import Test.HsBindgen.Runtime.CEnum qualified
 import Test.HsBindgen.Runtime.ConstantArray qualified
-import Test.HsBindgen.Runtime.HighLevel.RefineSmoke qualified
+import Test.HsBindgen.Runtime.HighLevel.CoreDump qualified
+import Test.HsBindgen.Runtime.HighLevel.Test qualified
 import Test.HsBindgen.Runtime.IncompleteArray qualified
 import Test.HsBindgen.Runtime.SizedByteArray qualified
 
@@ -20,7 +21,8 @@ main = defaultMain $ testGroup "test-runtime" [
     , Test.HsBindgen.Runtime.CBool.tests
     , Test.HsBindgen.Runtime.CEnum.tests
     , Test.HsBindgen.Runtime.ConstantArray.tests
-    , Test.HsBindgen.Runtime.HighLevel.RefineSmoke.tests
+    , Test.HsBindgen.Runtime.HighLevel.CoreDump.tests
+    , Test.HsBindgen.Runtime.HighLevel.Test.tests
     , Test.HsBindgen.Runtime.IncompleteArray.tests
     , Test.HsBindgen.Runtime.SizedByteArray.tests
     ]